RS/requirements.txt
torch
torchvision
pillow
spotipy
pandas
